+++ extension/src/helpers/urls.ts
@@ -3,12 +3,12 @@ import browser from "webextension-polyfill";
 import { TransactionInfo } from "../types/transactions";
 
 export interface BlobToSign {
-  isDomainListedAllowed: boolean,
-  domain: string,
-  tab: browser.Tabs.Tab | undefined,
-  blob: string,
-  url: string,
-  accountToSign: string
+  isDomainListedAllowed: boolean;
+  domain: string;
+  tab: browser.Tabs.Tab | undefined;
+  blob: string;
+  url: string;
+  accountToSign: string;
 }
 
 export const encodeObject = (obj: {}) =>
@@ -22,7 +22,9 @@ export const newTabHref = (path = "", queryParams = "") =>
 
 export const removeQueryParam = (url = "") => url.replace(/\?(.*)/, "");
 
-export const parsedSearchParam = (param: string): TransactionInfo | BlobToSign => {
+export const parsedSearchParam = (
+  param: string,
+): TransactionInfo | BlobToSign => {
   const decodedSearchParam = decodeString(param.replace("?", ""));
   return decodedSearchParam ? JSON.parse(decodedSearchParam) : {};
 };
