+++ extension/src/popup/ducks/access.ts
@@ -4,7 +4,7 @@ import {
   rejectAccess as internalRejectAccess,
   grantAccess as internalGrantAccess,
   signTransaction as internalSignTransaction,
-  signBlob as internalSignBlob
+  signBlob as internalSignBlob,
 } from "@shared/api/internal";
 
 export const grantAccess = createAsyncThunk("grantAccess", internalGrantAccess);
@@ -19,10 +19,7 @@ export const signTransaction = createAsyncThunk(
   internalSignTransaction,
 );
 
-export const signBlob = createAsyncThunk(
-  "signBlob",
-  internalSignBlob,
-);
+export const signBlob = createAsyncThunk("signBlob", internalSignBlob);
 
 // Basically an alias for metrics purposes
 export const rejectTransaction = createAsyncThunk(
@@ -31,8 +28,4 @@ export const rejectTransaction = createAsyncThunk(
 );
 
 // Basically an alias for metrics purposes
-export const rejectBlob = createAsyncThunk(
-  "rejectBlob",
-  internalRejectAccess,
-);
-
+export const rejectBlob = createAsyncThunk("rejectBlob", internalRejectAccess);
