+++ extension/src/popup/views/SignTransaction/styles.scss
@@ -1,4 +1,4 @@
-.SignTransaction {
+.SignTransaction, .SignBlob {
   height: var(--popup--height);
   overflow: hidden;
   position: relative;
